department and faculty added successfully

+++ src/main/java/org/parul/pmp/dto/FacultyDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class FacultyDTO {
 
-    private int department_id;
+    private long department_id;
     private String faculty_code;
     private String faculty_firstname;
     private String faculty_middlename;
@@ -26,11 +26,11 @@ public class FacultyDTO {
         this.departmentDTOS = departmentDTOS;
     }*/
 
-    public int getDepartment_id() {
+    public long getDepartment_id() {
         return department_id;
     }
 
-    public void setDepartment_id(int department_id) {
+    public void setDepartment_id(long department_id) {
         this.department_id = department_id;
     }
